cancel apply group

+++ app/controllers/groups_controller.rb
@@ -2,7 +2,8 @@ class GroupsController < ApplicationController
   #before_action :set_group, only: %i[ show edit update destroy ]
   before_action :set_group, except: %i[ index new create ]
   before_action :authenticate_user!
-  load_and_authorize_resource except: %i[ index new create apply_group join_group leave_group ]
+  #load_and_authorize_resource except: %i[ index new create apply_group join_group leave_group ]
+  load_and_authorize_resource except: %i[ index new create apply_group cancel_apply leave_group ]
   # 이렇게 해야 그룹에 속하지 않은 user가 새로 group을 생성할 수 있음.
   # 가입/탈퇴도 포함되어야 함...
 
@@ -33,15 +34,11 @@ class GroupsController < ApplicationController
   def create
     @group = Group.new(group_params)
 
-    # @group = current_user.groups.new(group_params)
-    # @group[:user_id] = current_user.id
-    # group에 user_id 항목이 없으니 굳이 이렇게 생성할 필요 없음.
-
     if Group.find_by(name: @group.name).present?
       flash.now[:danger] = "That group name has already been taken."
+
       render 'new'
     else
-
       @group.save
       # usergroup 먼저 save 불가. 아마도 group.id 때문이 아닐까?
 
@@ -50,14 +47,14 @@ class GroupsController < ApplicationController
       usergroup.group_id = @group.id
       usergroup.state = "active"
       usergroup.save
+      # 근데 이렇게 하면 group은 save에 성공하고 usergroup save에 실패했을 때 오류처리를 어떻게 하나
+      # group과 usergroup save를 한 트랜젝션에서 처리하고 싶은데...
 
       current_user.add_role :group_manager, @group
       current_user.add_role :group_member, @group
       # 생성한 그룹에 대해 group_manager role 부여
 
-      respond_to do |format|
-        format.html { redirect_to @group, notice: "Group was successfully created." }
-      end
+      redirect_to @group, notice: "Group was successfully created."
     end
   end
 
@@ -85,7 +82,7 @@ class GroupsController < ApplicationController
 
     # current_user.remove_role :group_member, @group
     # current_user.remove_role :group_manager, @group
-    # 그룹 권한 삭제 : 할 필요 없음 dependent: destroy
+    # 그룹 권한 삭제 : 굳이 할 필요 없음 rolify에서 dependent: destroy로 처리하는 거 같음.
 
     @group.destroy
     respond_to do |format|
@@ -93,20 +90,22 @@ class GroupsController < ApplicationController
     end
   end
 
-#  def join_group
-#    # approve 없이 그룹에 join
-#    if UserGroup.where(user_id: current_user.id, group_id: @group.id).count <= 0
-#      usergroup = UserGroup.new
-#      usergroup.user_id = current_user.id
-#      usergroup.group_id = @group.id
-#      usergroup.save
-#
-#      current_user.add_role :group_member, @group
-#      redirect_to @group, notice: "You've been added to #{@group.name}."
-#    else
-#      redirect_to groups_path, notice: "You're already a member of #{@group.name}."
-#    end
-#  end
+  def join_group
+    # approve 없이 그룹에 join
+    # 정책에 따라서 이렇게 허가 없이 join만 해도 바로 승인되는 경우를 허용해야 함...
+    if UserGroup.where(user_id: current_user.id, group_id: @group.id).count <= 0
+      usergroup = UserGroup.new
+      usergroup.user_id = current_user.id
+      usergroup.group_id = @group.id
+      usergroup.state = 'active'
+      usergroup.save
+
+      current_user.add_role :group_member, @group
+      redirect_to @group, notice: "You've been added to #{@group.name}."
+    else
+      redirect_to groups_path, notice: "You're already a member of #{@group.name}."
+    end
+  end
 
   def apply_group
     if UserGroup.where(user_id: current_user.id, group_id: @group.id).count <= 0
@@ -115,10 +114,21 @@ class GroupsController < ApplicationController
       usergroup.group_id = @group.id
       usergroup.state = 'pending'
       usergroup.save
-
       redirect_to @group, notice: "You've been applied for #{@group.name}."
     else
       redirect_to groups_path, notice: "You're already a member of #{@group.name}."
+    end
+  end
+
+  def cancel_apply
+    apply_user = User.find(params[:apply_user_id])
+
+    usergroup = apply_user.user_groups.find_by_group_id(@group.id)
+    if usergroup.state == 'pending' && usergroup.user_id == apply_user.id && usergroup.group_id == @group.id
+      usergroup.destroy
+      redirect_to groups_path, notice: "Applying has been canceled."
+    else
+      redirect_to @group, notice: "user/group 오류 또는 active 상태가 아님"
     end
   end
 
@@ -127,28 +137,35 @@ class GroupsController < ApplicationController
 
     if current_user.has_role? :group_manager, @group
       #usergroup = UserGroup.where(user_id: params[:apply_user_id], group_id: @group.id)
-      #이건 안 되는디
+      #이건 안 되는디(user_groups)
       usergroup = apply_user.user_groups.find_by_group_id(@group.id)
-      #이건 된다...왜????
+      #이건 된다...왜????(user_group)
 
-      if usergroup.state == 'pending'
+      if usergroup.state == 'pending' && usergroup.user_id == apply_user.id && usergroup.group_id == @group.id
         usergroup.state = 'active'
         usergroup.save
         apply_user.add_role :group_member, @group
         redirect_to @group, notice: "#{apply_user.name}'s been approved."
       else
-        redirect_to groups_path, notice: "뭔지는 모르지만 오류!"
+        redirect_to groups_path, notice: "user/group 오류 또는 pending 상태가 아님"
       end
     end
   end
 
   def leave_group
     if !current_user.has_role? :group_manager, @group
-      current_user.remove_role :group_member, @group
-      usergroup = current_user.user_groups.find_by_group_id(@group.id)
-      usergroup.destroy
-      current_user.save!
-      redirect_to groups_path, notice: "You're no longer a member of #{@group.name}."
+      if current_user.has_role? :group_member, @group
+        current_user.remove_role :group_member, @group
+        usergroup = current_user.user_groups.find_by_group_id(@group.id)
+        if usergroup.state == 'active' && usergroup.user_id == current_user.id && usergroup.group_id == @group.id
+          usergroup.destroy
+          redirect_to groups_path, notice: "You're no longer a member of #{@group.name}."
+        else
+          redirect_to @group, notice: "user/group 오류 또는 active 상태가 아님"
+        end
+      else
+        redirect_to @group, notice: "group_member가 아님"
+      end
     else
       redirect_to @group, notice: "group_manager는 탈퇴 불가"
     end
